+++ crates/router/src/core/payments/operations/payment_cancel.rs
@@ -158,6 +158,7 @@ impl<F: Send + Clone, Ctx: PaymentMethodRetrieve>
             setup_mandate: None,
             customer_acceptance: None,
             token: None,
+            token_data: None,
             address: core_types::PaymentAddress::new(
                 shipping_address.as_ref().map(From::from),
                 billing_address.as_ref().map(From::from),
@@ -165,6 +166,7 @@ impl<F: Send + Clone, Ctx: PaymentMethodRetrieve>
             ),
             confirm: None,
             payment_method_data: None,
+            payment_method_info: None,
             force_sync: None,
             refunds: vec![],
             disputes: vec![],
@@ -173,6 +175,7 @@ impl<F: Send + Clone, Ctx: PaymentMethodRetrieve>
             card_cvc: None,
             creds_identifier,
             pm_token: None,
+            payment_method_status: None,
             connector_customer_id: None,
             recurring_mandate_payment_data: None,
             ephemeral_key: None,