+++ crates/router/src/types/api/verify_connector.rs
@@ -88,6 +88,7 @@ impl VerifyConnectorData {
             payment_method_token: None,
             connector_api_version: None,
             recurring_mandate_payment_data: None,
+            payment_method_status: None,
             connector_request_reference_id: attempt_id,
             address: types::PaymentAddress::new(None, None, None),
             payment_id: common_utils::generate_id_with_default_len(